2nd Commit-Full Project (not complete but working)
I updated the models for Message, CastingCall, Image, And User. I updated the repositories. Some features such as Username in post and message are missing, as well as some queries in the Repositories. I have yet to work out what I need to add. I am submitting this for peer review